Refactor log handling by creating useLogActions for better organization and maintainability
